`QueryResponse()` is the centerpiece of the library
`QueryResponse()` takes a byte array and returns a byte array. It's a
black box that `main.go` can use to input the DNS query and get back the
DNS response. This enables us to have a very lean `main.go`, which means
we can put much of the processing into the library, and which means we
can unit-test the components.

- A better-late-than-never `gofmt -w .` included cosmetic changes.