🐞 make sslip.io's website accessible again
The new webserver didn't have A and AAAA records for `sslip.io`, which
meant there was no website. We need a website. This commit fixes that.

The code is somewhat inflexible in that it assumes that there's exactly
oneA record and exactly one AAAA record.

I took the opportunity to bump SOA's serial and set the Hostmaster email
address to my primary email (brian.cunnie@gmail.com).

fixes:
```
curl: (6) Could not resolve host: sslip.io
```