🐞 Dockerfile: cunnie/sslip.io-dns-server → 1.2.2
Rather than using Docker Hub's automated build feature (which doesn't
seem to work when setting up new repositories), I've opted to manually
build & push the images.

There are workarounds which might allow me to use GitHub's automated
build feature, like creating an organization, moving the repos to the
new organization, and creating a 'bot' user to publish the images, but
that seems like a lot of work for little gain.

fixes:

> Fetch source repositories failed.

> Connect a GitHub account to cunnie to enable automated builds. If it is already connected, please re-link the source provider.