`_acme-challenge.` TXT records return NS not SOA
This change is to enable wildcard certificates via DNS challenge.

My earlier attempt failed: when queried for TXT for
`_acme-challenge.127-0-0-1.sslip.io`, I returned an authoritative
response with the Authorities section containing the SOA record, which
signaled, "There is no TXT record, of that I am sure." Even though I had
configured an NS record `_acme-challenge.127-0-0-1.sslip.io` to return
`127-0-0-1.sslip.io`.

Now I return an authoritative response with an NS record, not an SOA
record, in the response.

But that's still not enough, and I'd like to do the following changes:

- When queried for a DNS-01 challenge, e.g.
`_acme-challenge.127-0-0-1.sslip.io`, I return a _non-authoritative_
response with the NS record. This is the behavior of, say, `dig ns
sslip.io @a0.nic.io.`

- When queried for any NS records, I return an Additionals section
with the IP addresses.