Delete `pipeline-simple.yml`; it's old
This pipeline's only purpose was an asset in a blog post that I wrote a
couple of years ago, and is no longer necessary.

Also, and this sounds petty, but I didn't like the RED on my CI--I'd
like to see as much green as possible. Now my CI is green (with the
exception of the many-colored "badges" pipeline).