Use builtin key-value store when no etcd
We use a simple map as a key-value store when there's no `etcd`.

We test both: when there's etcd, and when there isn't.

I'm inordinately pleased at my intuitive understanding of
`DescribeTable()`, allowing me to define a function and using it in the
two places it's tested (with `etcd` and without), which is much better
than having duplicate copies of the same test (requiring twice the
maintenance when updating tests).

TODO: Don't run `etcd` unit tests when there's no `etcd`