The Xip struct is constant, not volatile
Prior to this commit, the Xip struct served two masters: global state
(e.g. metrics) and volatile state (querier's source IP address).

This was ugly, but workable.

But with the advent of the blocklist it became untenable. I needed the
Xip struct to be truly global, to download only one copy of the
blocklist, not one copy for each of the network interfaces that
sslip.io-dns-server was listening to. Hence this change.

One the downside I had to plumb the querier's source IP address through
several layers of function calls.