Accept `-port` flag to bind to ports other than 53
We want to allow users to bind to ports other than 53. A big reason is
that port 53 is a privileged port, and often requires root privileges.
We don't want to force our users to use root privileges in order to run
the tests.

This isn't a problem on macOS, but is on Linux.