Laying the groundwork for passed-in configuration
The massive 80+ line `Customizations` variable is a hard-coded
monstrosity, and I've fallen out of love with it.

I'd like the customizations to be passed in from the caller, in this
case, `main.go`.

To that end, I've created a `default.json`, which should contain all the
customizations with the exception of the key-value functionality, which
I don't have a good way to deal with just yet.