Dockerfile pipeline: build using HEAD, not tag
This change allows me to build new versions of the Docker images without
forcing me to create a gratuitous release merely for that purpose.

In this specific case, I had neglected to have `curl` follow redirects
(`-L`), thus the Docker images created were missing the sslip.io DNS
server.

On the downside, it's possible that the *nginx* Docker containers are
built with content/features that haven't been released yet, but I'm
willing to run that risk.