Test TCP binding failure modes
- If it can't bind to all addresses via TCP, log the ones it could &
  couldn't bind to & keep running
- If it can't bind to any address via TCP, keep running (unlike UDP
  which must fail)

The big challenge in writing these tests is that the binding behavior is
different for macOS (Ventura 13.6 (22G120)) than for Linux.
Specifically, to "squat" on an address, macOS must listen on ALL TCP
addresses (INADDR_ANY) plus the specific address. Linux only needs to
listen to the specific address.

I have no idea what the behavior on Windows is.

I also removed listenPort as a top-level variable; it didn't need to be
top level.