"ns-azure is dead, long live ns-hetzner"
The nameserver on Azure is probably my least-favorite: much slower, much
higher latency. Even though it would've made more geographic sense to
dismantle my GCP nameserver in favor of the Hetzner, I'm using this
opportunity to get rid of the Azure.

And, of course, introduce the Hetzner nameserver with its 20TB of
bandwidth allowance, which I've come to need.