Run rspec tests locally, easily
We test our four production nameservers using `rspec`, and although that
worked fine when we used GitHub actions with a curated Ruby, it didn't
always work as well when running locally.

With this commit we introduce a Gemfile which has the necessary Ruby
dependencies, which make it easer to run the tests locally.