🐞 "undefined method 'success?' for nil"
When placating the linter, substituting `$CHILD_STATUS` for `$?`, I
broke the tests. The linter neglected to say that this substitution
needed a `require "English"`. This commit brings in "English".

This sort of linting is overreach in my opinion: `$?` is universal for
people who have coded bash scripts, and is easily learned for those who
haven't. `$CHILD_STATUS` is not universal, only meaningful for English
speakers, and has additional dependencies.

This is a learning experience: I shouldn't commit & push code that I
can't test because I'm on the train and their internet hijacks the DNS
service.

Thanks <https://github.com/rubocop/rubocop/issues/1747>