Retire ns-gce.sslip.io nameserver
The ns-gce nameserver has always been problematic: I couldn't turn on
logging because Google gouged me >$20/month for the logs alone (which
was the impetus for the `-quiet` flag), so I made it "half" a
nameserver: the registrar didn't have it as one of the three
nameservers, but the nameservers themselves, when queried for the NS
records, included it as a fourth.

But I want complete logs, so I'm fully retiring this nameserver.

Unfortunately, I signed up for a three-year GCP "committed use discount"
which doesn't expire until 2026-11-02, which goes to show that a lot can
happen in two years. Who knew that VMware would deprecate k8s in favor
of Cloud Foundry? Who knew that I'd leave Broadcom for Majestic Labs in
July 2024? Lesson learned: don't do 3-year committed use discounts.